requirements.txt
Django==3.0.7
Pillow==7.1.2
django-admin-sortable2==0.7.6
django-tinymce==3.0.2
python-dotenv==0.9.1



